backend/pyproject.toml
[project]
name = "aris"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.15.1",
    "beautifulsoup4>=4.13.3",
    "fastapi[standard]>=0.115.11",
    "lxml>=5.3.2",
    "psycopg2-binary>=2.9.10",
    "rsm-markup>=0.3",
    "sqlalchemy>=2.0.39",
    "ujson>=5.10.0",
    "uvicorn>=0.34.0",
    "python-jose[cryptography]>=3.4.0",
    "pydantic-settings>=2.9.1",
    "pydantic[email]>=2.11.1",
    "bcrypt>=4.3.0",
    "python-dotenv>=1.1.0",
    "asyncpg>=0.30.0",
    "greenlet>=3.1.1",
]


[tool.setuptools]
package-dir = {"" = "."}


[tool.ruff]
extend-select = ["I"]    # Enable import sorting
line-length = 100


[tool.pylsp.plugins.pycodestyle]
max_line_length = 100


[tool.pylsp.plugins.flake8]
enabled = false
